Logical Ops: Move logical operations to separate ops file
